Revert "Create v2 management API"
