+++ conf.d/management_apis.conf
@@ -33,30 +33,23 @@ server {
     access_log /var/log/api-gateway/access.log main;
     error_log /var/log/api-gateway/management.log debug;
 
-    location ~ /(?<version>v2)/(?<tenantId>[^/]*)/apis/?(?<api_id>[^/]*)?/?(?<query>[^/]*)? {
+    location ~ /v1/apis/?(?<api_id>[^/]*)?/?(?<query>[^/]*)? {
         access_by_lua_block {
-            local apis = require("management/routes/apis")
-            apis.requestHandler()
-        }
-    }
-
-    location ~ /(?<version>v1)/apis/?(?<api_id>[^/]*)?/?(?<query>[^/]*)? {
-        access_by_lua_block {
-            local apis = require("management/routes/apis")
+            local apis = require("management/apis")
             apis.requestHandler()
         }
     }
 
     location ~ /v1/tenants/?(?<tenant_id>[^/]*)?/?(?<query>[^/]*)? {
         access_by_lua_block {
-            local tenants = require("management/routes/tenants")
+            local tenants = require("management/tenants")
             tenants.requestHandler()
         }
     }
 
     location /v1/subscriptions {
          access_by_lua_block {
-             local subscriptions = require("management/routes/subscriptions")
+             local subscriptions = require("management/subscriptions")
              subscriptions.requestHandler()
          }
     }
