+++ scripts/lua/management/resources.lua
