+++ tests/install-deps.sh
@@ -11,4 +11,3 @@ luarocks install --tree=lua_modules luasocket
 luarocks install --tree=lua_modules sha1
 luarocks install --tree=lua_modules md5
 luarocks install --tree=lua_modules net-url
-luarocks install --tree=lua_modules luafilesystem
