--- tests/scripts/lua/management/examples/example1.json
@@ -1,86 +1,0 @@
-{
-  "swagger": "2.0",
-  "info": {
-    "version": "1.0",
-    "title": "Hello World API"
-  },
-  "basePath": "/native",
-  "schemes": [
-    "https"
-  ],
-  "consumes": [
-    "application/json"
-  ],
-  "produces": [
-    "application/json"
-  ],
-  "paths": {
-    "/hello": {
-      "get": {
-        "description": "Returns a greeting to the user!",
-        "parameters": [
-          {
-            "name": "user",
-            "in": "path",
-            "type": "string",
-            "required": true,
-            "description": "The name of the user to greet."
-          }
-        ],
-        "responses": {
-          "200": {
-            "description": "Returns the greeting.",
-            "schema": {
-              "type": "string"
-            }
-          },
-          "400": {
-            "description": "Invalid characters in \"user\" were provided."
-          }
-        }
-      }
-    }
-  },
-  "securityDefinitions": {
-    "client_id": {
-      "type": "apiKey",
-      "name": "X-Api-Key",
-      "in": "header"
-    },
-    "google": {
-      "type": "oauth2",
-      "flow": "tokenIntrospect",
-      "x-tokenintrospect": {
-        "url": "https://www.googleapis.com/oauth2/v3/tokeninfo"
-      }
-    }
-  },
-  "security": [
-    {
-      "client_id": [],
-      "google": []
-    }
-  ],
-  "x-gateway-rate-limit": [
-    {
-      "unit": "minute",
-      "units": 3,
-      "rate": 100
-    }
-  ],
-  "x-gateway-configuration": {
-    "cors": {
-      "enabled": true
-    },
-    "assembly": {
-      "execute": [
-        {
-          "proxy": {
-            "target-url": "https://appconnect.mybluemix.net",
-            "verb": "keep"
-          }
-        }
-      ]
-    }
-  }
-}
